modified:   SimpleGeometry/src/main/AndroidManifest.xml 	modified:   SimpleGeometry/src/main/java/uk/ks/jarvis/simple/geometry/fragments/CreateFigureDialog.java 	modified:   SimpleGeometry/src/main/java/uk/ks/jarvis/simple/geometry/fragments/ShapeDialog.java 	modified:   SimpleGeometry/src/main/java/uk/ks/jarvis/simple/geometry/holders/BaseHolder.java 	modified:   SimpleGeometry/src/main/java/uk/ks/jarvis/simple/geometry/shapes/Circle.java 	modified:   SimpleGeometry/src/main/java/uk/ks/jarvis/simple/geometry/shapes/Dot.java 	modified:   SimpleGeometry/src/main/java/uk/ks/jarvis/simple/geometry/shapes/Line.java 	modified:   SimpleGeometry/src/main/java/uk/ks/jarvis/simple/geometry/shapes/Shape.java 	modified:   SimpleGeometry/src/main/java/uk/ks/jarvis/simple/geometry/shapes/ShapeList.java 	modified:   SimpleGeometry/src/main/java/uk/ks/jarvis/simple/geometry/utils/BaseHelper.java 	modified:   SimpleGeometry/src/main/res/layout/create_figure_dialog.xml 	modified:   SimpleGeometry/src/main/res/layout/dot_options_fragment.xml

+++ SimpleGeometry/src/main/java/uk/ks/jarvis/simple/geometry/utils/BaseHelper.java
@@ -51,10 +51,12 @@ public class BaseHelper {
         point1.setY(point2.getY());
     }
 
-    public static float getAngleFrom2Points(Point firstPoint, Point secondPoint) {
+    public static double getAngleFrom2Points(Point firstPoint, Point secondPoint) {
         Point coord = new Point(abs(secondPoint.getX() - firstPoint.getX()), abs(secondPoint.getY() - firstPoint.getY()));
         double tgOfAngle = coord.getY()/coord.getX();
         double angle = arctg(tgOfAngle);
+//                Point coord = new Point(abs(secondPoint.getX() - firstPoint.getX()), abs(secondPoint.getY() - firstPoint.getY()));
+//                double angle = (90 / (coord.getX() + coord.getY())) * coord.getY();
 
         if ((secondPoint.getY() > firstPoint.getY()) && (secondPoint.getX() > firstPoint.getX())) { // if 1 coordinate plane
 
@@ -66,6 +68,6 @@ public class BaseHelper {
             angle = (360 - angle);
         }
         angle = 360 - (angle % 360);
-        return (float) angle;
+        return angle;
     }
 }
